+++ bookwyrm/templates/manage_invites.html
@@ -27,7 +27,7 @@
 <div class="block">
     <h2 class="title is-4">Generate New Invite</h2>
 
-    <form name="invite" action="/create-invite/" method="post">
+    <form name="invite" action="/invite/" method="post">
         {% csrf_token %}
         <div class="field is-grouped">
             <div class="control">
