+++ bookwyrm/views/__init__.py
@@ -1,3 +1,4 @@
 ''' make sure all our nice views are available '''
 from .authentication import Login, Register, Logout
 from .password import PasswordResetRequest, PasswordReset, ChangePassword
+from .invite import ManageInvites, Invite
